# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Forge Core API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List modules
      responses:
        "200":
          description: Modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [modules]
      summary: Create module
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
        "400":
          description: Invalid manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /modules/{id}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get module with pipeline state
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateModule
      tags: [modules]
      summary: Update module
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
        "403":
          description: Blocked (locked core impact)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteModule
      tags: [modules]
      summary: Delete module
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /modules/{id}/activate:
    post:
      operationId: activateModule
      tags: [modules]
      summary: Activate module
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /modules/{id}/deactivate:
    post:
      operationId: deactivateModule
      tags: [modules]
      summary: Deactivate module
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /modules/{id}/install:
    post:
      operationId: installModule
      tags: [modules]
      summary: Request module installation (runs Governor decision pipeline)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Governor decision for this install request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernorDecision"
        "403":
          description: Install blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /modules/{id}/rollback:
    post:
      operationId: rollbackModule
      tags: [modules]
      summary: Roll module back to previous snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Rolled back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
        "400":
          description: No snapshot available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /modules/{id}/snapshots:
    get:
      operationId: listModuleSnapshots
      tags: [modules]
      summary: List module snapshots
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModuleSnapshot"
  /modules/{id}/guardian-review:
    post:
      operationId: runGuardianReview
      tags: [guardian]
      summary: Run Guardian review on a module
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Guardian review report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuardianReview"
  /guardian-reviews:
    get:
      operationId: listGuardianReviews
      tags: [guardian]
      summary: List Guardian reviews
      parameters:
        - name: moduleId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Guardian reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuardianReview"
  /governor-decisions:
    get:
      operationId: listGovernorDecisions
      tags: [governor]
      summary: List Governor decisions
      parameters:
        - name: moduleId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Governor decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GovernorDecision"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Module:
      type: object
      required: [id, name, type, version, status, riskLevel, testStatus, installStatus, ownerAgent, active, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        type:
          type: string
          enum:
            - planner
            - architect
            - frontend-builder
            - backend-builder
            - database-builder
            - test-generator
            - security-reviewer
            - ux-reviewer
            - documentation-generator
            - performance-analyzer
            - dependency-checker
        purpose: { type: ["string", "null"] }
        version: { type: string }
        status: { type: string }
        active: { type: boolean }
        dependencies:
          type: array
          items: { type: string }
        ownerAgent: { type: string }
        riskLevel:
          type: string
          enum: [low, medium, high]
        testStatus: { type: string }
        installStatus: { type: string }
        rollbackInfo: { type: ["string", "null"] }
        manifest: { type: ["string", "null"] }
        touchesCore: { type: boolean }
        createdAt: { type: string }
    ModuleDetail:
      type: object
      required: [module, testRuns, guardianReviews, governorDecisions, approvals, snapshots]
      properties:
        module:
          $ref: "#/components/schemas/Module"
        testRuns:
          type: array
          items:
            $ref: "#/components/schemas/TestRun"
        guardianReviews:
          type: array
          items:
            $ref: "#/components/schemas/GuardianReview"
        governorDecisions:
          type: array
          items:
            $ref: "#/components/schemas/GovernorDecision"
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/Approval"
        snapshots:
          type: array
          items:
            $ref: "#/components/schemas/ModuleSnapshot"
    ModuleInput:
      type: object
      required: [name, type]
      properties:
        name: { type: string, minLength: 1 }
        type:
          type: string
          enum:
            - planner
            - architect
            - frontend-builder
            - backend-builder
            - database-builder
            - test-generator
            - security-reviewer
            - ux-reviewer
            - documentation-generator
            - performance-analyzer
            - dependency-checker
        purpose: { type: string }
        version: { type: string }
        dependencies:
          type: array
          items: { type: string }
        ownerAgent: { type: string }
        riskLevel:
          type: string
          enum: [low, medium, high]
        manifest: { type: string }
        touchesCore: { type: boolean }
    ModuleUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        purpose: { type: string }
        version: { type: string }
        dependencies:
          type: array
          items: { type: string }
        ownerAgent: { type: string }
        riskLevel:
          type: string
          enum: [low, medium, high]
        manifest: { type: string }
        touchesCore: { type: boolean }
    ModuleSnapshot:
      type: object
      required: [id, moduleId, version, createdAt]
      properties:
        id: { type: integer }
        moduleId: { type: integer }
        version: { type: string }
        data: { type: ["string", "null"] }
        reason: { type: ["string", "null"] }
        createdAt: { type: string }
...
    TestRun:
      type: object
      required: [id, status, types, createdAt]
      properties:
        id: { type: integer }
        moduleId: { type: ["integer", "null"] }
        sandboxId: { type: ["integer", "null"] }
        types:
          type: array
          items: { type: string }
        status: { type: string }
        results: { type: ["string", "null"] }
        passed: { type: ["integer", "null"] }
        failed: { type: ["integer", "null"] }
        createdAt: { type: string }
...
    Approval:
      type: object
      required: [id, moduleId, level, status, createdAt]
      properties:
        id: { type: integer }
        moduleId: { type: integer }
        moduleName: { type: ["string", "null"] }
        level:
          type: string
          enum: [auto, review, owner, blocked]
        status: { type: string }
        reason: { type: ["string", "null"] }
        decidedBy: { type: ["string", "null"] }
        decidedAt: { type: ["string", "null"] }
        createdAt: { type: string }
...
    GuardianReview:
      type: object
      required: [id, moduleId, outcome, findings, createdAt]
      properties:
        id: { type: integer }
        moduleId: { type: integer }
        moduleName: { type: ["string", "null"] }
        outcome:
          type: string
          enum: [pass, warning, fail]
        findings:
          type: array
          items:
            $ref: "#/components/schemas/GuardianFinding"
        createdAt: { type: string }
    GuardianFinding:
      type: object
      required: [category, severity, message]
      properties:
        category: { type: string }
        severity: { type: string }
        message: { type: string }
    GovernorDecision:
      type: object
      required: [id, moduleId, decision, rationale, createdAt]
      properties:
        id: { type: integer }
        moduleId: { type: integer }
        moduleName: { type: ["string", "null"] }
        decision:
          type: string
          enum: [install_allowed, install_blocked, review_required, rollback_required]
        rationale: { type: string }
        inputs: { type: ["string", "null"] }
        createdAt: { type: string }
...